# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for analyzing and searching relevant Reddit content based on topics",
        "title": "Reddit Content Analyzer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/v1/reddit/relevance/search": {
            "post": {
                "description": "Searches Reddit posts based on a topic and returns posts that are relevant according to the specified criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reddit"
                ],
                "summary": "Search for relevant Reddit posts",
                "parameters": [
                    {
                        "description": "Search request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.RelevanceRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with relevant posts",
                        "schema": {
                            "$ref": "#/definitions/github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.RelevanceResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.RelevanceRequestDto": {
            "type": "object",
            "required": [
                "search_method",
                "topic"
            ],
            "properties": {
                "created_after": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "min_num_comments": {
                    "type": "integer"
                },
                "relevance_threshold": {
                    "type": "number"
                },
                "search_method": {
                    "enum": [
                        "search",
                        "latest"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.SearchMethod"
                        }
                    ]
                },
                "subreddits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.RelevanceResponseDto": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.SubRedditPostDto"
                    }
                }
            }
        },
        "github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.SearchMethod": {
            "type": "string",
            "enum": [
                "search",
                "latest"
            ],
            "x-enum-varnames": [
                "SearchMethodSearch",
                "SearchMethodLatest"
            ]
        },
        "github_com_ReyOrtiz_reddit-content-analyzer_internal_contracts.SubRedditPostDto": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "is_relevant": {
                    "type": "boolean"
                },
                "num_comments": {
                    "type": "integer"
                },
                "relevance_score": {
                    "type": "number"
                },
                "relevance_summary": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "subreddit_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}